+++ src/services/IdentityNowClient.ts
@@ -6,7 +6,7 @@ import { withQuery } from "../utils/UriUtils";
 import { compareByName, convertToText, isEmpty } from "../utils";
 import { DEFAULT_ACCOUNTS_QUERY_PARAMS } from "../models/Account";
 import { DEFAULT_ENTITLEMENTS_QUERY_PARAMS } from "../models/Entitlements";
-import { Configuration, IdentityProfilesApi, IdentityProfile, LifecycleState, LifecycleStatesApi, Paginator, ServiceDeskIntegrationApi, ServiceDeskIntegrationDto, Source, SourcesApi, Transform, TransformsApi, WorkflowsBetaApi, WorkflowBeta, WorkflowExecutionBeta, WorkflowLibraryTriggerBeta, ConnectorRuleManagementBetaApi, ConnectorRuleResponseBeta, ConnectorRuleValidationResponseBeta, AccountsApi, AccountsApiListAccountsRequest, Account, EntitlementsBetaApi, EntitlementsBetaApiListEntitlementsRequest, PublicIdentitiesApi, PublicIdentitiesApiGetPublicIdentitiesRequest, Entitlement, PublicIdentity, JsonPatchOperationBeta, SPConfigBetaApi, SpConfigImportResultsBeta, SpConfigJobBeta, ImportOptionsBeta, SpConfigExportResultsBeta, ObjectExportImportOptionsBeta, ExportPayloadBetaIncludeTypesEnum, ImportSpConfigRequestBeta } from 'sailpoint-api-client';
+import { Configuration, IdentityProfilesApi, IdentityProfile, LifecycleState, LifecycleStatesApi, Paginator, ServiceDeskIntegrationApi, ServiceDeskIntegrationDto, Source, SourcesApi, Transform, TransformsApi, WorkflowsBetaApi, WorkflowBeta, WorkflowExecutionBeta, WorkflowLibraryTriggerBeta, ConnectorRuleManagementBetaApi, ConnectorRuleResponseBeta, ConnectorRuleValidationResponseBeta, AccountsApi, AccountsApiListAccountsRequest, Account, EntitlementsBetaApi, EntitlementsBetaApiListEntitlementsRequest, PublicIdentitiesApi, PublicIdentitiesApiGetPublicIdentitiesRequest, Entitlement, PublicIdentity, JsonPatchOperationBeta, SPConfigBetaApi, SpConfigImportResultsBeta, SpConfigJobBeta, ImportOptionsBeta, SpConfigExportResultsBeta, ObjectExportImportOptionsBeta, ExportPayloadBetaIncludeTypesEnum, ImportSpConfigRequestBeta, TransformRead } from 'sailpoint-api-client';
 import { DEFAULT_PUBLIC_IDENTITIES_QUERY_PARAMS } from '../models/PublicIdentity';
 import axios, { AxiosInstance, AxiosResponse } from 'axios';
 import { ImportEntitlementsResult } from '../models/JobStatus';
@@ -194,18 +194,7 @@ export class IdentityNowClient {
 		}
 		console.log("resetSource: endpoint = " + endpoint);
 		const httpClient = await this.getAxios(CONTENT_TYPE_FORM_URLENCODED);
-		const response = await httpClient.post(endpoint)
-			.catch(error => {
-				console.error(error);
-				if (error.response) {
-					let detail = error.response.data;
-					if (error.response.status === 400) {
-						detail = detail.exception_message;
-					}
-					throw new Error("Could not reset source: " + detail);
-				}
-				throw error;
-			});
+		const response = await httpClient.post(endpoint);
 		return response.data;
 	}
 
@@ -273,7 +262,7 @@ export class IdentityNowClient {
 	 * It will return sorted by name list by name
 	 * @returns all transforms of the tenant
 	 */
-	public async getTransforms(): Promise<Transform[]> {
+	public async getTransforms(): Promise<TransformRead[]> {
 		console.log("> getTransforms");
 		const apiConfig = await this.getApiConfiguration();
 		const api = new TransformsApi(apiConfig);
@@ -285,7 +274,7 @@ export class IdentityNowClient {
 		return transforms;
 	}
 
-	public async getTransformByName(name: string): Promise<Transform> {
+	public async getTransformByName(name: string): Promise<TransformRead> {
 		console.log("> getTransformByName", name);
 		const apiConfig = await this.getApiConfiguration();
 		const api = new TransformsApi(apiConfig);